file/fn renamed + url check added
